+++ src/wrappers.rs
@@ -5,6 +5,7 @@ pub mod mj_statistic;
 pub mod mj_rendering;
 pub mod mj_auxiliary;
 pub mod mj_primitive;
+pub mod mj_editing;
 pub mod mj_option;
 pub mod mj_model;
 pub mod mj_data;
@@ -12,7 +13,10 @@ pub mod mj_data;
 pub mod fun;
 
 pub use mj_visualization::*;
+pub use mj_statistic::*;
 pub use mj_rendering::*;
+pub use mj_auxiliary::*;
+pub use mj_primitive::*;
 pub use mj_option::*;
 pub use mj_model::*;
 pub use mj_data::*;