+++ src/wrappers/mj_rendering.rs
@@ -127,7 +127,9 @@ impl MjrContext {
         unsafe { mjr_setAux(index as i32, self.ffi_mut()); }
     }
 
-    /// Draws a text overlay.
+    /// Draws a text overlay. The optional `overlay2` parameter displays additional overlay, next to `overlay`.
+    /// # Panics
+    /// When the `overlay` or `overlay2` contain '\0' characters, a panic occurs.
     pub fn overlay(&mut self, font: MjtFont, gridpos: MjtGridPos, viewport: MjrRectangle, overlay: &str, overlay2: Option<&str>) {
         let c_overlay = CString::new(overlay).unwrap();
         let c_overlay2 = overlay2.map(|x| CString::new(x).unwrap());