+++ docs/guide/source/changelog.rst
@@ -14,6 +14,16 @@ This project uses `semantic versioning <https://semver.org/>`_.
 
 Unreleased
 ================================
+- :ref:`model_editing` support, which can be used to procedurally generate |mj_model|. It can be used
+  for terrain generation, parameter randomization, etc. 
+
+  - Added module :docs-rs:`mujoco_rs::wrappers::mj_editing`.
+  - Added two examples. One on basic model editing and one on terrain generation.
+
+- :docs-rs:`~mujoco_rs::viewer::<struct>MjViewer` and :docs-rs:`~mujoco_rs::renderer::<struct>MjRenderer`:
+
+  - Increased the headroom for visual-only geoms, which aren't drawn by the user, from 100 to 2000.
+
 - :docs-rs:`~mujoco_rs::viewer::<struct>MjViewer`:
 
   - Added events for keys:
@@ -31,7 +41,6 @@ Unreleased
       - ``Z``: light,
       - ``T``: transparent,
       - ``I``: inertia.
-
 
 - :docs-rs:`~mujoco_rs::wrappers::mj_visualization::<type>MjvCamera`: